fix(cpu/matmul): widen GEMV weight row offsets to long (#429)
Two GEMV weight-offset expressions computed a row offset with int arithmetic,
so a single tensor past the wrap point produced a negative offset and an
out-of-bounds read rather than a clean failure.

ComputeRows (`weightsQ8 + row * rowBytes`, 7 sites across the AVX-512+VNNI,
AVX-512, AVX2 and scalar tiers) is the reachable one: the stride is in bytes
(~1.0625 per weight), so the product wraps at 2 GB of tensor. Llama 3.1 405B's
LM head — vocab 128256, hidden 16384, rowBytes (16384/32)*34 = 17408 — gives
128256 * 17408 = 2,232,680,448, over int.MaxValue by ~85 MB. The first affected
row is 123,362, so the last 4,894 rows (3.8% of vocab) read from a wrapped
negative offset on a shipping model.

GemvF16 (`weightsHalf + row * k`, 2 sites) counts elements, not bytes, so on
that same model m*k = 2,101,346,304 stays ~2.2% under the limit. Widened for
consistency, but it needs a tensor larger than any current model to trigger.

Only the GEMV paths were exposed. ComputeGemmTiled clamps row < tileRows <= 256
and pre-offsets its base pointer via `(long)mStart * q8RowBytes`, so the tiled
path was already safe and is untouched. Activation-buffer offsets are left
alone: each indexes a buffer whose size is the same product, so wrapping would
need an >8 GB logits allocation.

This matches the widening convention already used by GemmF16, GemmF32,
ComputeGemmTiled and the F16 tiled row read — the Gemm paths had the cast, the
Gemv paths did not.

No performance cost. Verified by diffing the JIT's x64 for GemvQ8_0 between
builds: the baseline already sign-extended the product (`imul ecx,r12d` then
`movsxd rcx,ecx`); the cast moves the widening ahead of the multiply
(`imul rcx,rax` on an already-widened row), which is what fixes the overflow.
The 4-row fast path is one instruction shorter; total method size 1024 -> 1028
bytes. The arithmetic is in the outer row loop, amortized over 16-128 VNNI
block iterations per row. A position-balanced timing harness could not resolve
any delta: within-arm spread was 46-88%, dwarfing all measured differences.

Reported-by: unsafePtr

+++ src/DotLLM.Cpu/Kernels/MatMul.cs
@@ -138,7 +138,7 @@ public static unsafe partial class MatMul
             for (; row + 3 < m; row += 4)
             {
                 VecDotQ8_0Vnni_4Rows(
-                    weightsQ8 + row * rowBytes,
+                    weightsQ8 + (long)row * rowBytes,
                     weightsQ8 + (row + 1) * rowBytes,
                     weightsQ8 + (row + 2) * rowBytes,
                     weightsQ8 + (row + 3) * rowBytes,
@@ -146,7 +146,7 @@ public static unsafe partial class MatMul
             }
             for (; row < m; row++)
             {
-                result[row] = VecDotQ8_0Avx512(weightsQ8 + row * rowBytes, xQ8, blockCount);
+                result[row] = VecDotQ8_0Avx512(weightsQ8 + (long)row * rowBytes, xQ8, blockCount);
             }
         }
         else if (Avx512BW.IsSupported)
@@ -156,7 +156,7 @@ public static unsafe partial class MatMul
             for (; row + 3 < m; row += 4)
             {
                 VecDotQ8_0Avx512_4Rows(
-                    weightsQ8 + row * rowBytes,
+                    weightsQ8 + (long)row * rowBytes,
                     weightsQ8 + (row + 1) * rowBytes,
                     weightsQ8 + (row + 2) * rowBytes,
                     weightsQ8 + (row + 3) * rowBytes,
@@ -164,7 +164,7 @@ public static unsafe partial class MatMul
             }
             for (; row < m; row++)
             {
-                result[row] = VecDotQ8_0Avx512(weightsQ8 + row * rowBytes, xQ8, blockCount);
+                result[row] = VecDotQ8_0Avx512(weightsQ8 + (long)row * rowBytes, xQ8, blockCount);
             }
         }
         else if (Avx2.IsSupported)
@@ -174,7 +174,7 @@ public static unsafe partial class MatMul
             for (; row + 3 < m; row += 4)
             {
                 VecDotQ8_0Avx2_4Rows(
-                    weightsQ8 + row * rowBytes,
+                    weightsQ8 + (long)row * rowBytes,
                     weightsQ8 + (row + 1) * rowBytes,
                     weightsQ8 + (row + 2) * rowBytes,
                     weightsQ8 + (row + 3) * rowBytes,
@@ -182,14 +182,14 @@ public static unsafe partial class MatMul
             }
             for (; row < m; row++)
             {
-                result[row] = VecDotQ8_0Avx2(weightsQ8 + row * rowBytes, xQ8, blockCount);
+                result[row] = VecDotQ8_0Avx2(weightsQ8 + (long)row * rowBytes, xQ8, blockCount);
             }
         }
         else
         {
             for (int row = 0; row < m; row++)
             {
-                result[row] = VecDotQ8_0Scalar(weightsQ8 + row * rowBytes, xQ8, blockCount);
+                result[row] = VecDotQ8_0Scalar(weightsQ8 + (long)row * rowBytes, xQ8, blockCount);
             }
         }
     }
@@ -1564,7 +1564,7 @@ public static unsafe partial class MatMul
             float* rowBuf = stackalloc float[k];
             for (int row = 0; row < m; row++)
             {
-                var srcRow = new ReadOnlySpan<Half>(weightsHalf + row * k, k);
+                var srcRow = new ReadOnlySpan<Half>(weightsHalf + (long)row * k, k);
                 var destRow = new Span<float>(rowBuf, k);
                 TensorPrimitives.ConvertToSingle(srcRow, destRow);
                 y[row] = TensorPrimitives.Dot(destRow, new ReadOnlySpan<float>(x, k));
@@ -1577,7 +1577,7 @@ public static unsafe partial class MatMul
             {
                 for (int row = 0; row < m; row++)
                 {
-                    var srcRow = new ReadOnlySpan<Half>(weightsHalf + row * k, k);
+                    var srcRow = new ReadOnlySpan<Half>(weightsHalf + (long)row * k, k);
                     var destRow = rented.AsSpan(0, k);
                     TensorPrimitives.ConvertToSingle(srcRow, destRow);
                     y[row] = TensorPrimitives.Dot(destRow, new ReadOnlySpan<float>(x, k));
